Delete the legacy array-of-structs event store and relocate its differential oracle into the tests
