+++ tests/Unit/EventLogExpert.Runtime.Tests/LogTable/CombinedColumnViewDifferentialTests.cs
@@ -5,13 +5,14 @@ using EventLogExpert.Eventing.Common.Channels;
 using EventLogExpert.Eventing.Common.EventLogs;
 using EventLogExpert.Eventing.Common.Events;
 using EventLogExpert.Runtime.LogTable;
+using EventLogExpert.Runtime.Tests.LogTable.TestSupport;
 using EventLogExpert.Runtime.Tests.TestUtils;
 
 namespace EventLogExpert.Runtime.Tests.LogTable;
 
 /// <summary>
 ///     Differential parity for <see cref="CombinedColumnView" /> (the live SoA multi-log backing) against the
-///     retained AoS <see cref="CombinedEventView" /> oracle, locking the column-direct K-way merge for the operations
+///     relocated AoS <c>ReferenceCombinedView</c> oracle, locking the column-direct K-way merge for the operations
 ///     selection and the highlight cache depend on: <c>Count</c>, <c>LocatorAt</c>, <c>Rank</c>, <c>ResolveByKey</c>, and
 ///     a re-merge after a per-log append.
 /// </summary>
@@ -26,7 +27,7 @@ public sealed class CombinedColumnViewDifferentialTests
     private static readonly string[] s_tasks = ["", "Logon", "Service"];
 
     [Fact]
-    public void CombinedColumnView_MatchesCombinedEventViewOracle_AfterPerLogReMerge()
+    public void CombinedColumnView_MatchesReferenceOracle_AfterPerLogReMerge()
     {
         // Design SS14 test 15: appending to one log and re-merging must keep the combined column view aligned with the
         // AoS oracle (both Rank/LocatorAt/ResolveByKey and merged order) at the same and grown corpus.
@@ -59,7 +60,7 @@ public sealed class CombinedColumnViewDifferentialTests
     [InlineData(42)]
     [InlineData(2026)]
     [InlineData(8675309)]
-    public void CombinedColumnView_MatchesCombinedEventViewOracle_OverRandomContexts(int seed)
+    public void CombinedColumnView_MatchesReferenceOracle_OverRandomContexts(int seed)
     {
         var corpus = new Corpus(seed, logCount: 2 + (seed % 3));
         corpus.SeedInterleaved(totalEvents: 300);
@@ -79,7 +80,7 @@ public sealed class CombinedColumnViewDifferentialTests
         }
     }
 
-    private static void AssertParity(CombinedColumnView column, CombinedEventView oracle, Corpus corpus, string label)
+    private static void AssertParity(CombinedColumnView column, ReferenceCombinedView oracle, Corpus corpus, string label)
     {
         Assert.Equal(oracle.Count, column.Count);
 
@@ -169,18 +170,16 @@ public sealed class CombinedColumnViewDifferentialTests
             for (int i = 0; i < count; i++) { AppendOne(logIndex); }
         }
 
-        public (CombinedColumnView Column, CombinedEventView Oracle) BuildViews(SortContext context)
+        public (CombinedColumnView Column, ReferenceCombinedView Oracle) BuildViews(SortContext context)
         {
             var columnViews = new List<EventColumnView>(_perLog.Count);
-            var oracleLists = new List<SegmentedSortedList>(_perLog.Count);
 
             for (int k = 0; k < _perLog.Count; k++)
             {
                 columnViews.Add(DisplayViewTestFactory.Build(_logIds[k], _perLog[k], context));
-                oracleLists.Add(SegmentedSortedList.CreateSorted(_perLog[k], context));
             }
 
-            return (new CombinedColumnView(columnViews, context), new CombinedEventView(oracleLists, context));
+            return (new CombinedColumnView(columnViews, context), new ReferenceCombinedView(_perLog, context));
         }
 
         public IEnumerable<ResolvedEvent> SampleEvents()
@@ -198,8 +197,8 @@ public sealed class CombinedColumnViewDifferentialTests
 
         public void SeedInterleaved(int totalEvents)
         {
-            // Guarantee every log is non-empty (CombinedEventView keys on list[0].OwningLog), then interleave the rest
-            // at random so the logs weave together in the merged stream.
+            // Guarantee every log is non-empty so every log participates in the merge, then interleave the rest at random
+            // so the logs weave together in the merged stream.
             for (int k = 0; k < _perLog.Count; k++) { AppendOne(k); }
 
             for (int i = _perLog.Count; i < totalEvents; i++) { AppendOne(_rng.Next(_perLog.Count)); }
@@ -219,6 +218,90 @@ public sealed class CombinedColumnViewDifferentialTests
                 TaskCategory = s_tasks[_rng.Next(s_tasks.Length)],
                 ComputerName = s_computers[_rng.Next(s_computers.Length)]
             });
+        }
+    }
+
+    // Test-only re-implementation of the deleted AoS CombinedEventView oracle over a reference-sorted flattened list. The
+    // corpus is totally ordered (distinct OwningLog per log, globally unique monotonic clock, per-log RecordId), so the
+    // K-way merge the production oracle performed collapses to sorting the concatenated events by the reference comparer;
+    // Rank/ResolveByKey reproduce the oracle's reference-or-value-key window match verbatim.
+    private sealed class ReferenceCombinedView
+    {
+        private readonly Comparison<ResolvedEvent> _comparer;
+        private readonly ResolvedEvent[] _merged;
+
+        public ReferenceCombinedView(IReadOnlyList<IReadOnlyList<ResolvedEvent>> perLog, SortContext context)
+        {
+            var flattened = new List<ResolvedEvent>();
+
+            foreach (var log in perLog) { flattened.AddRange(log); }
+
+            _comparer = AosReferenceOrdering.Reference(
+                context.OrderBy, context.IsDescending, context.GroupBy, context.IsGroupDescending);
+
+            int[] order = AosReferenceOrdering.Order(
+                flattened, context.OrderBy, context.IsDescending, context.GroupBy, context.IsGroupDescending);
+            _merged = new ResolvedEvent[order.Length];
+
+            for (int i = 0; i < order.Length; i++) { _merged[i] = flattened[order[i]]; }
+        }
+
+        public int Count => _merged.Length;
+
+        public ResolvedEvent this[int index] => _merged[index];
+
+        public int Rank(ResolvedEvent target)
+        {
+            ArgumentNullException.ThrowIfNull(target);
+
+            for (int i = LowerBound(target); i < _merged.Length && _comparer(_merged[i], target) == 0; i++)
+            {
+                if (Matches(_merged[i], target)) { return i; }
+            }
+
+            return -1;
+        }
+
+        public ResolvedEvent? ResolveByKey(ResolvedEvent candidate)
+        {
+            ArgumentNullException.ThrowIfNull(candidate);
+
+            for (int i = LowerBound(candidate); i < _merged.Length && _comparer(_merged[i], candidate) == 0; i++)
+            {
+                if (Matches(_merged[i], candidate)) { return _merged[i]; }
+            }
+
+            return null;
+        }
+
+        // Reference first, then the (RecordId + time + OwningLog + LogName) value key; a null RecordId never matches by key
+        // so distinct error-read events stay separate - identical to the deleted oracle's window scan.
+        private static bool Matches(ResolvedEvent current, ResolvedEvent target)
+        {
+            if (ReferenceEquals(current, target)) { return true; }
+
+            if (current.RecordId is null || target.RecordId is null) { return false; }
+
+            return current.RecordId == target.RecordId
+                && current.TimeCreated == target.TimeCreated
+                && string.Equals(current.OwningLog, target.OwningLog, StringComparison.Ordinal)
+                && string.Equals(current.LogName, target.LogName, StringComparison.Ordinal);
+        }
+
+        private int LowerBound(ResolvedEvent target)
+        {
+            int low = 0;
+            int high = _merged.Length;
+
+            while (low < high)
+            {
+                int mid = low + ((high - low) >> 1);
+
+                if (_comparer(_merged[mid], target) < 0) { low = mid + 1; }
+                else { high = mid; }
+            }
+
+            return low;
         }
     }
 }
